Modify test scripts to take the path to where the game is installed.

+++ test/Air_SE/test_all.sh
@@ -2,6 +2,12 @@
 
 # Runs all paths of AIR; the result should be a save file with everything
 # unlocked.
+if [ ! -n "$1" ]
+then
+  echo "Usage: `basename $0` <path to AIR directory>"
+  exit 65
+fi
+GAMEDIR=$1
 
 rm -Rf ~/.rlvm/KEY_AIR_SE/
 mkdir -p AIR.log
@@ -11,7 +17,7 @@ function runPath() {
   LOG=`echo $SCRIPT | sed s/\.lua/\.log/g;`
 
   echo "Running $SCRIPT..."
-  time build/luaRlvm --count-undefined test/Air_SE/$SCRIPT /storage/GameImages/AIR/ > AIR.log/$LOG 2>&1
+  time build/luaRlvm --count-undefined test/Air_SE/$SCRIPT $GAMEDIR > AIR.log/$LOG 2>&1
 }
 
 runPath "Dream_Kano.lua"
