+++ test/P_BRIDE/test_all.sh
@@ -2,6 +2,12 @@
 
 # Runs all paths of Princess Bride!; the result should be a save file with
 # everything unlocked.
+if [ ! -n "$1" ]
+then
+  echo "Usage: `basename $0` <path to P_BRIDE directory>"
+  exit 65
+fi
+GAMEDIR=$1
 
 rm -Rf ~/.rlvm/130CM_PRINCESS_BRIDE_SE/
 mkdir -p P_BRIDE.log
@@ -11,7 +17,7 @@ function runPath() {
   LOG=`echo $SCRIPT | sed s/\.lua/\.log/g;`
 
   echo "Running $SCRIPT..."
-  time build/luaRlvm --count-undefined test/P_BRIDE/$SCRIPT /storage/GameImages/P_BRIDE > P_BRIDE.log/$LOG 2>&1
+  time build/luaRlvm --count-undefined test/P_BRIDE/$SCRIPT $GAMEDIR > P_BRIDE.log/$LOG 2>&1
 }
 
 runPath "Kiyomi.lua"
